an alternative to c.eval
!!! This commit represents a robustness and performance improvement, but
it's not on a critical path, so we can ignore it for now !!!

Instead of relying on eval, we now expect pickle formats to instantiate
themselves for us at compile-time.

This makes it irrelevant whether the format implicit is imported like
"import json._" or declared locally - everything will work in both cases
(the latter would fail when doing eval). Also this provides a performance
improvement, because eval is very costly.