moves pickling macros into bundles
the true purpose of this change will become clear from the
subsequent commit. for now it can be viewed as cleanup.

this implies moving them into a separate file. this is done to:
on the one hand, avoid https://issues.scala-lang.org/browse/SI-5954.
on the other hand, to not pollute package.scala (interfaces) with
macro code which becomes more and more fat (and will become even bigger)