moves some logic from picklers to pickle formats
I propose viewing picklers as hollow shells, which only provide infrastructure
for pickling and unpickling of stuff.

Why? Because picklers generally lack knowledge to do anything useful.
For example, they can't even deal with nulls - this has to be
delegated to pickle formats.

This once again changes the interface of PickleFormat. Now it's the same
as the interface of Pickler, but shifted one metalevel up.